Hooked the event system to the WindowsWindow class
The "Application" is now aware that a RGF event has occurred. This is because the event system is now "hooked" to the "WindowsWindow" class and now recieves events from GLFW. Application now has a "OnEvent" function that will receieve all RGF events. The window now closes because application has created a "OnWindowClose" callback that closes the window.