Fixed some minor bugs. See desc
Some if not all nivida cards could not use the batch rendering shader, so this should be fixed for now.
^^Because of this^^ --> Changed the max texture unit slots from what the GPU provides to hard limit of 32.

The engine actually shutdown properly now. This was because not all the rendering data was cleared and
the engine was in a infinite loop.

--- bugs known ---
For some reason, even though openal compiles and is apart of the engine, sometimes the audio engine will init properly on other machines.
It compiles on every device i've tested but I get reports that openal cannot init sometimes.
^^ Will investigate ^^