Now scenes reference assets via their Handle. see desc
- Right now textures are the only asset in pixelate.
- If you were to save or load a scene the asset handle is serialized.
- The asset manager is used to load an asset already in its table when loading a scene.

this is still early on and a few bugs and issues are in the engine. For example moving assets will cause them to not load correctly because the filesystem is not aware that the
asset has moved.

More assets such as audio, shaders, animation files and more will be supported via this system.