Cached shaders WIP see desc
There seems to be an opengl driver bug on my AMD card that prevents me from using glshaderbinary properly. I still cache the shaders to disk and load the file if they exist but instead we generate shader sources as a workaround. This should be tested on other devices to see if this bug still occurs