Can start to draw material properties -- see desc
The sprite renderer component gui is being redone. We only draw floats for now but we can draw almost all types...BUT WE CANNOT MODIFY SAMPLED IMAGES IN THE EDITOR.

This is because sampled images are not allowed in UBOs in vulkan and thus it is not reflected properly while using SPIRV.
Also, the renderer expects default shader attributes too. This means that at some point the renderer needs to be changed to allow any type of material instead of expecting a texture.
To solve this though, we need to make sure we can reflect sampled images properly first. Also the fact that opengl expects the attributes to be stored on the vao is kinda annoying.......this needs to be rethought when trying to make the attribues dynamic