Default white textures work now...see desc
The material system works! But it's missing a few editor pieces that need to be done, such as editing in the material edtior and not the entity sprite component.

Also the renderer is kinda at odds with itself... we kinda dont use 32 texture arrays anymore. I don't even think we need to index the default texture at index 0 anymore and just render using materials. Since we sorta simulate this via our render pass system and therefore do not need texture arrays.
I'm not sure on what the performance implicataion is by doing this but it seems stable enough and the performance hasn't seem to tanked as of yet.....